Fix #43: Allow to insert icons into empty paragraphs
Currently, it is not possible to insert Font Awesome icons into empty
paragraphs; instead this results in the paragraph being removed. This
is actually caused by tinymce/tinymce#3175, but as the issue might not
be resolved in the near future, we apply the workaround of additionally
inserting a non-breaking space after the icon.

This solution doesn't work on old IE (IE <= 8) as these browsers don't
have the `Node.textContent` property. A fallback for such browsers
could be implemented by using `Node::innerText` instead, but that
might not be necessary due to the decreasing usage of IE 8.